BSCN/frontend/requirements.txt
streamlit>=1.26.0
requests>=2.28
python-docx>=0.8.11
PyPDF2>=3.0.0
reportlab>=4.0
